sshproxy-managerd: ignore SIGPIPE events
They are generated by systemd when journald is restarted and
sshproxy-managerd is running under systemd, logging to stdout and not
restarted.

We also switch logging to syslog. With systemd >= 219 it should not be
necessary but Red Hat reverted the needed functionality from the systemd
used on RHEL 7.x:

Patch:  0125-Revert-journald-allow-restarting-journald-without-lo.patch
Reason: https://lists.freedesktop.org/archives/systemd-devel/2015-February/028685.html