change how we store connections in etcd
WARNING: this commit breaks how connections are stored in etcd.
Administrators should remove old entries before using this new version.

Before this commit, connections were aggregated in etcd and we only
store the number of connections from a user and the time of last
connection. The number was incremented when a connection was made and
decremented when sshproxy exited.

We noticed two problems:
  - sometimes the number was not decremented,
  - sshproxy crashed when exiting but without any harm (it was just
  annoying in the logs).

With this commit, each connection is individually stored in etcd as a
key with a configurable time to live. The key will be kept alive until
sshproxy exits. The key will then disappear from etcd after its time to
live.

sshproxyctl was changed to display either aggregated connections (the
default) or all connections.